% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapCI.R
\name{bootstrapCI}
\alias{bootstrapCI}
\title{Determine confidence intervals to rainfall intensities estimated by a PDF function.}
\usage{
bootstrapCI(Intensity, Parameters, Type = "Gumbel", Rsample = 1000,
  Return.P, Conf.Inter = 0.95)
}
\arguments{
\item{Intensity}{a numeric vector with intensity [mm/h] values of different 
years for a specific time duration (\emph{e.g.} 5, 15, 120 minutes, \emph{etc.})}

\item{Parameters}{list with three elements: (i) type of distribution function 
(ii) fitted parameters, and (iii) source to call specfic function in the \pkg{lmomco} package.}

\item{Type}{a character specifying a name of the probability distribution function fitted 
(see \code{\link{selecDIST}}) by \code{\link{fitDISTRI}} function.}

\item{Rsample}{An integer representing number of resamples to conduct when 
confidence interval will be computed.}

\item{Return.P}{a numeric vector with return periods like non-exceedance probabilities.}

\item{Conf.Inter}{level of the confidence interval.}
}
\value{
A list of:

 \itemize{
  \item \code{nonexceed.prob} a numeric vector with non-exceedance probabilities.
  \item \code{lower.lim} a numeric vector confidence bound lower for quantile estimates.
  \item \code{upper.lim} a numeric vector confidence bound upper for quantile estimates.
  \item \code{Para.set} a matrix containing estimated distribution parameters for each resample.
  \item \code{quantiles} a matrix of quantile estimates for each resample.
 }
}
\description{
Conducts bootstrap to randomly sample of intensity values 'n' times for a 
specified distribution to estimate the confidence interval for each given 
non-exceedance probability.
}
\examples{

data(inten)
data(Pargumbel)
Tp <- c(2, 3, 5, 10, 25, 50, 100)
FR <- 1 - 1/Tp
CI.test <- bootstrapCI(Intensity = inten, Parameters = Pargumbel, Type = "gumbel",
Rsample = 50, Return.P =FR, Conf.Inter = 0.90)
(CI.test$Conf.Inter[,2])
# result = 125.9501, 139.6132, 154.6232, 173.6736, 197.5024, 215.0634, 232.4939
 
}
\author{
David Zamora <dazamoraa@unal.edu.co> 
Water Resources Engineering Research Group - GIREH
}
