% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fractures_crit.R
\docType{data}
\name{fractures.crit}
\alias{fractures.crit}
\title{fractures.crit}
\format{A matrix with four rows and five variables:
\describe{
  \item{Info. Criteria}{AIC, BIC, AICc and KIC, in dimensionless units}   
  \item{PDFs}{Normal, Exponential, Lognormal, Gamma and GEV}
  ...
}}
\usage{
fractures.crit
}
\description{
A dataset containing the values of four information criterias calculated 
five probability density functions.
}
\keyword{datasets}

