% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goodFIT.R
\name{goodFIT}
\alias{goodFIT}
\title{Title}
\usage{
goodFIT(Station = ..., Type = ..., Intensity = ..., Parameters = ...,
  M.fit = ..., Dura = .., Plot = ..., Resolution = 300, SAVE = FALSE)
}
\arguments{
\item{Station:}{a character specifying a name or number of pluviographic 
station where data were measurement, and it use to save results in *.xls 
format.}

\item{Type:}{a character specifying a name of probability distribution 
function fitted (see \code{\link{selecDIST}}) by \code{\link{fitDISTRI}} function.}

\item{Intensity:}{a numeric vector with intensity values for a specific time 
duration in different return periods.}

\item{Parameters:}{a list with three elements: i) Type of distribution function ii)
Parameters fitted, and iii) source to call specfic function in the \pkg{lmomco} package.}

\item{M.fit:}{a character specifying a name of fit method employed on pdf, just three 
options are available: L-moments (\emph{Lmoments}), Probability-Weighted Moments (\emph{PWD}), 
and Maximum Likelihood (\emph{MLE}).}

\item{Dura:}{a character specifying a time duration of the \code{Intensity}, (e.g. 30 min). 
This parameter is used to save results.}

\item{Plot:}{a number (1) to determine if it will be plotted density curves both empirical 
as modeled (\emph{pdf}). If you use other number the graphs will not appear.}

\item{Resolution:}{a number to determine resolution that the plot function used to save graphs. 
It can have two options: 300 and 600 ppi. See \code{\link{resoPLOT}}.}

\item{SAVE:}{a logical value. TRUE will save \code{Plot} but if is FALSE just show \code{Plot}.}
}
\value{
A data frame with statistics values of goodness of fit tests and its respective p-value, 
moreover information criteria are evaluated:

\itemize{
 \item \emph{Kolmogorov-Smirnov}: statistic= KS and p-value1
 \item \emph{Anderson-Darling}: statistic= AD and p-value2
 \item \emph{Cramer-von Mises}: statistic= Omega2 and p-value3
 \item \emph{Akaike's Information Criterion}: AIC
 \item \emph{Bayesian Information Criterion}: BIC
}
}
\description{
This function computing of goodness-of-fit for continuous univariate 
distributions using tests: Kolmogorov-Smirnov, Anderson-Darling and 
Cramer-von Mises, it based on \pkg{goftest} package. Moreover information criteria 
are evaluated: Akaike's Information Criterion and Bayesian Information 
Criterion by means of \code{\link{InfoCRIT}} function.
}
\examples{
# Meteorology station in the Airport Farfan in Tulua, Colombia.
data(Intgum5min)
data(Pagumbel)
# not plotted
test.fit <- goodFIT(Station = "2610516", Type = "Gumbel", Intensity = Intgum5min,
                    Parameters = Pargumbel,M.fit = "Lmoments", Dura ="5_min", Plot = 0)
 
}
\author{
David Zamora <dazamoraa@unal.edu.co> 
Water Resources Engineering Research Group - GIREH
}

