% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IDFCurve.R
\name{IDFCurve}
\alias{IDFCurve}
\title{IDFCurve}
\usage{
IDFCurve(Data = ..., Station = "2610516", Duration = FALSE,
  Periods = FALSE, Type = "gumbel", M.fit = "lmoments", Plot = 1234,
  Strategy = 1:3, logaxe = "", CI = FALSE, iter = 500,
  goodtest = FALSE, Resolution = 300, SAVE = FALSE)
}
\arguments{
\item{Data:}{1}

\item{Station:}{11}

\item{Duration:}{11}

\item{Periods:}{11}

\item{Type:}{11}

\item{M.fit:}{11}

\item{Plot:}{11}

\item{Strategy:}{11}

\item{logaxe:}{11}

\item{CI:}{11}

\item{iter:}{11}

\item{goodtest:}{11}

\item{Resolution:}{11}

\item{SAVE:}{11}
}
\value{
A list of:

 \itemize{
   \item \code{Parameters} a list
   \item \code{Int.pdf}
   \item \code{Conf.Inter}
   \item \code{goodness.fit}
   \item \code{Info.PDF}
 }
}
\description{
IDFCurve
}
\examples{

}
\author{
David Zamora <dazamoraa@unal.edu.co>
Water Resources Engineering Research Group - GIREH
}

