% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IDFdata.R
\docType{data}
\name{IDFdata}
\alias{IDFdata}
\title{IDFdata}
\format{A matrix with eight rows and seven columns:
\describe{
  \item{Duration}{5, 10, 15, 20, 30, 60 120 and 360, in minutes}
  \item{return periods}{2, 3, 5, 10, 25, 50 and 100, in years}
  \item{Rainfall intensity}{numeric, millimiter/hour}
  ...
}}
\usage{
IDFdata
}
\description{
A numeric matrix with values of rainfall intensities to different duration times and
calculated for seven return periods. This values were obtained from processed data of the
meteorology station in the Farfan Airport in Tulua, Colombia.
}
\keyword{datasets}

