% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InforCRIT.R
\name{InfoCRIT}
\alias{InfoCRIT}
\title{InfoCRIT}
\usage{
InfoCRIT(Intensity = ..., Parameters = ...)
}
\arguments{
\item{Intensity:}{numeric vector with intensity values for a duration 
specific in different return periods.}

\item{Parameters:}{list with three elements: i) type of distribution 
function ii) parameters fitted, and iii) source to call specfic function
in the package \pkg{lmomco}.}
}
\value{
Provided a numeric vector with the corresponding AIC and BIC values.
}
\description{
This function determine values of the criteria Akaike's Information Criterion
(AIC) and Bayesian Information Criterion to fitted probability density 
functions with lmomco package
}
\examples{
# Meteorology station in the Airport Farfan in Tulua, Colombia. 
data(Intgum5min)
data(Pargumbel)
InfoCRIT(Intensity=Intgum5min,Parameters=Pargumbel)
# Result: AIC= 86.73 & BIC= 86.62

}
\author{
David Zamora <dazamoraa@unal.edu.co> 
Water Resources Engineering Research Group - GIREH
}
