% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inten.R
\docType{data}
\name{inten}
\alias{inten}
\title{inten}
\format{A numeric matrix of rainfall intensities with 35 rows and 9 columns:
\describe{
  \item{Duration times}{5, 10, 15, 20, 30, 120 and 360, in minutes}
  \item{Years}{Record period, 1972 to 2010}
  \item{Rainfall intensity}{Intensity, in millimeter/hour}
  ...
}}
\usage{
inten
}
\description{
A dataset with rainfall intensities for eight duration times which were 
measured in meteorology station at the Farfan Airport in Tulua, Colombia, 
during period of 1972 to 2010.
}
\keyword{datasets}
