% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Intgum5min.R
\docType{data}
\name{Intgum5min}
\alias{Intgum5min}
\title{Intgum5min}
\format{A numeric vector of rainfall intensities with 7 values:
\describe{
  \item{Duration times}{5 in minutes}
  \item{Return periods}{Record period, 2, 3, 5, 10, 25, 50 and 100 in years}
  \item{Rainfall intensity}{Intensity in millimeter/hour}
  ...
}}
\usage{
Intgum5min
}
\description{
A dataset with rainfall intensities (duration five (5) minutes) 
computed for seven return periods (i.e., 2, 3, 5, 10, 25, 50 and 100 years) 
using the \code{\link{IDFCurve}} function (fitted Gumbel-Lmomnts) 
in the meteorology station at the Farfan Airport in Tulua, 
Colombia, during period of 1972 to 2010.
}
\keyword{datasets}
