% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLEZ.R
\name{MLEZ}
\alias{MLEZ}
\title{MLEZ}
\usage{
MLEZ(Intensity, type, para.int = NULL, silent = TRUE,
  null.on.not.converge = TRUE, ptransf = function(t) return(t),
  pretransf = function(t) return(t), ...)
}
\arguments{
\item{...}{}

\item{Intensity:}{a numeric vector with intensity [mm/h] values of different
years for a specific time duration (\emph{e.g.} 5, 15, 120 minutes, \emph{etc}.).}

\item{type:}{a character specifying the name of distribution function that it will 
be employed: exponencial, gamma, gev, gumbel, log.normal3, normal, pearson, log.pearson3 and 
wakeby (see \code{\link{selecDIST}}).}

\item{silent:}{A logical to silence the \code{\link{try}}() function wrapping the \code{\link{DEoptim}}() function.}

\item{null.on.not.converge:}{A logical to trigging simple return of NULL if the optim() function returns a nonzero convergence status.}
}
\description{
This function fitting parameters of probability distribution functions (see \code{\link{selecDIST}})
by means of maximum likelihood and Performs evolutionary global optimization via the 
Differential Evolution algorithm (see \pkg{DEoptim} package).
}

