% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MME_DIST.r
\name{MME_DIST}
\alias{MME_DIST}
\title{Fitting parameters of PDF by means of moments method}
\usage{
MME_DIST(Intensity, Type)
}
\arguments{
\item{Intensity}{a numeric vector with intensity [mm/h] values of different
years for a specific time duration (\emph{e.g.} 5, 15, 120 minutes, \emph{etc}.).}

\item{Type}{a character specifying the name of distribution function that it will 
be employed: exponencial, gamma, gev, gumbel, log.normal3, normal, pearson, log.pearson3 and 
wakeby (see \code{\link{selecDIST}}).}
}
\value{
A list of:

 \itemize{
   \item \code{Parameters} a list with type of distribution fitted and values of its parameters
 }
}
\description{
This function allows to fit several distribution functions to observed data
by means of the moments method.
}
\examples{

# Meteorology station in the Farfan Airport in Tulua, Colombia.
data(inten)
Test.moments <- MME_DIST(Intensity = inten[,4], Type = "gumbel") 
## Results: xi = 71.50 ; alpha = 14.21

}
\author{
Albeiro Figueroa <cafigueroao@unal.edu.co> 
Water Resources Engineering Research Group - GIREH
}
