% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_criteria.R
\name{p.criteria}
\alias{p.criteria}
\title{pW Criteria}
\usage{
p.criteria(metrics = .., numod = .., pdfnames = ..)
}
\arguments{
\item{metrics:}{}

\item{numod:}{}

\item{pdfnames:}{}
}
\value{
Provided a intregral number of pW criteria and name of the best PDF
}
\description{
This function determine what probability distribution function has the best
goodness-of-fit to observations, number of parameters and the quality of 
model parameter estimates, taking in account different criterias information in a 
\emph{Siena et al., 2017}. \cite{(Siena et al., 2017)}.
}
\examples{
# Tales

}
\author{
Adriana Pina <appinaf@unal.edu.co> and David Zamora <dazamoraa@unal.edu.co> 
Water Resources Engineering Research Group - GIREH
}

