% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pargumbel.R
\docType{data}
\name{Pargumbel}
\alias{Pargumbel}
\title{Pargumbel}
\format{A list  with 3 elements according to structure of \pkg{lmomco} package:
\describe{
  \item{type PDF}{gum: Gumbel}
  \item{para}{xi and alpha parameters}
  \item{source}{A character using to call a \pkg{lmomco} functions}
  ...
}}
\usage{
Pargumbel
}
\description{
A list with Gumbel PDF parameters fitted by Lmoments method for rainfall intensities 
measured in meteorology station at the Farfan Airport in Tulua, Colombia, 
during period of 1972 to 2010.
}
\keyword{datasets}
