% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDFs.R
\name{pexp1}
\alias{pexp1}
\title{Cumulative probability of Exponential distribution}
\usage{
pexp1(q, xi, alpha)
}
\arguments{
\item{q:}{A real value vector.}

\item{xi:}{\eqn{\xi} is a location parameter (see \code{\link{parexp}} function).}

\item{alpha:}{\eqn{\alpha} s a scale parameter (see \code{\link{parexp}} function).}
}
\value{
Nonexceedance probability (\emph{F}) for q.
}
\description{
This function computes the cumulative probability or nonexceedance probability
of the Exponential distribution given parameters \eqn{\xi} and \eqn{\alpha}. Based in function 
of \pkg{lmomco} package.
}
\examples{

}
\author{
David Zamora <dazamoraa@unal.edu.co> 
Water Resources Engineering Research Group - GIREH
}

