% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDFs.R
\name{pgam1}
\alias{pgam1}
\title{Cumulative probability of Gamma distribution}
\usage{
pgam1(q, alpha, beta)
}
\arguments{
\item{q:}{a real value vector.}

\item{alpha:}{\eqn{\alpha} is a scale parameter (see \code{\link{pargam}} function).}

\item{beta:}{\eqn{\beta} is a location parameter (see \code{\link{pargam}} function).}
}
\value{
Nonexceedance probability (\emph{F}) for q.
}
\description{
This function computes the cumulative probability or nonexceedance probability 
of the Gamma distribution given parameters (\eqn{\alpha} and \eqn{\beta}) 
computed. Based in function of \pkg{lmomco} package.
}
\author{
David Zamora <dazamoraa@unal.edu.co> 
Water Resources Engineering Research Group - GIRE
}
