% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDFs.R
\name{pgev1}
\alias{pgev1}
\title{pgev1}
\usage{
pgev1(q, xi, alpha, kappa)
}
\arguments{
\item{q:}{a real value vector.}

\item{xi:}{\eqn{\xi} is a location parameter (see \code{\link{pargev}} function).}

\item{alpha:}{\eqn{\alpha} is a scale parameter (see \code{\link{pargev}} function).}

\item{kappa:}{\eqn{\kappa} is a shape parameter (see \code{\link{pargev}} function).}
}
\value{
Nonexceedance probability (\emph{F}) for q.
}
\description{
This function computes the cumulative probability or nonexceedance probability 
of the Generalized Extreme Value distribution given parameters (\eqn{\xi}, \eqn{\alpha} and \eqn{\kappa}). 
Based in function of \pkg{lmomco} package.
}
\examples{

}
\author{
David Zamora <dazamoraa@unal.edu.co> 
Water Resources Engineering Research Group - GIRE
}

