% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDFs.R
\name{pgum1}
\alias{pgum1}
\title{Cumulative probability of Gumbel distribution}
\usage{
pgum1(q, xi, alpha)
}
\arguments{
\item{q}{a real value vector.}

\item{xi}{\eqn{\xi} is a location parameter (see \code{\link{pargum}} function).}

\item{alpha}{\eqn{\alpha} is a scale paramete (see \code{\link{pargum}} function).}
}
\value{
Nonexceedance probability (\emph{F}) for q.
}
\description{
This function computes the cumulative probability or nonexceedance probability 
of the Gumbel distribution given parameters (\eqn{\xi} and \eqn{\alpha}). Based in function of 
lmomco package.
}
\author{
David Zamora <dazamoraa@unal.edu.co> 
Water Resources Engineering Research Group - GIRE
}
