% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDFs.R
\name{pln31}
\alias{pln31}
\title{Cumulative probability of LogNormal distribution}
\usage{
pln31(q, zeta, mulog, sigmalog)
}
\arguments{
\item{q}{a real value vector.}

\item{zeta}{\eqn{\zeta} is a lower bounds (see \code{\link{parln3}} function).}

\item{mulog}{\eqn{\mu} is a location parameter (see \code{\link{parln3}} function).}

\item{sigmalog}{\eqn{\sigma} is a scale parameter (see \code{\link{parln3}} function).}
}
\value{
Nonexceedance probability (\emph{F}) for q.
}
\description{
This function computes the cumulative probability or nonexceedance probability 
of the Log-Normal3 distribution given parameters (\eqn{\zeta}, \eqn{\mu} and \eqn{\sigma}). Based in function 
of \pkg{lmomco} package.
}
\author{
David Zamora <dazamoraa@unal.edu.co> 
Water Resources Engineering Research Group - GIRE
}
