% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDFs.R
\name{pnor1}
\alias{pnor1}
\title{Cumulative probability of Normal distribution}
\usage{
pnor1(q, mu, sigma)
}
\arguments{
\item{q:}{a real value vector.}

\item{mu:}{\eqn{\mu} is the arithmetic mean (see \code{\link{parnor}} function).}

\item{sigma:}{\eqn{\sigma} is the standard deviation (see \code{\link{parnor}} function).}
}
\value{
Nonexceedance probability (\emph{F}) for q.
}
\description{
This function computes the cumulative probability or nonexceedance 
probability of the Normal distribution given parameters (\eqn{\mu} and \eqn{\sigma}). 
Based in function of \pkg{lmomco} package.
}
\examples{

}
\author{
David Zamora <dazamoraa@unal.edu.co> 
Water Resources Engineering Research Group - GIRE
}

