% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDFs.R
\name{ppe31}
\alias{ppe31}
\title{ppe31}
\usage{
ppe31(q, mu, sigma, gamma)
}
\arguments{
\item{q:}{a real value vector.}

\item{mu:}{\eqn{\mu} is the mean (see \code{\link{parexp}} function).}

\item{sigma:}{\eqn{\sigma} is the standard deviation (see \code{\link{parexp}} function).}

\item{gamma:}{\eqn{\gamma} is the skew (see \code{\link{parexp}} function).}
}
\value{
Nonexceedance probability (\emph{F}) for q.
}
\description{
This function computes the cumulative probability or nonexceedance 
probability of the Pearson Type III distribution given parameters 
(\eqn{\mu}, \eqn{\sigma}, and \eqn{\gamma}). Based in function of \pkg{lmomco} package.
}
\examples{

}
\author{
David Zamora <dazamoraa@unal.edu.co> 
Water Resources Engineering Research Group - GIRE
}

