% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_criteria.R
\name{pW Criteria}
\alias{p.criteria}
\alias{pW Criteria}
\title{Identification criteria allow estimating the relative degree of likelihood 
of each model by means of their posterior weights (for AIC and AICc) or 
posterior model probability (for BIC and KIC)}
\usage{
p.criteria(metrics = .., critnames = .., pdfnames = ..)
}
\arguments{
\item{metrics:}{a numeric matrix with the values of criterias AIC, BIC, AICc and KIC (rows)
evaluated by each PDFs (columns).}

\item{critnames:}{a character vector with the names of information criteria evaluated.}

\item{pdfnames:}{a character vector with the names of PDFs evaluated.}
}
\value{
Provided a matrix with values of pW criteria by each PDFs and information criterias.
}
\description{
This function determine what probability distribution function has the best
goodness-of-fit to observations, number of parameters and the quality of 
taking in account different information criterias that are evaluated in 
the criteria proposed by (\emph{Siena et al., 2017}).
}
\examples{
# Example of five PDFs and their respectives values of four information criterias.
data(fractures.crit)

pW.1 <- p.criteria(metrics = fractures.crit, critnames = rownames(fractures.crit), pdfnames = colnames(fractures.crit))
pW.1
# GEV: AIC= 100, BIC= 100, AICc= 1, KIC= 99 
}
\author{
Adriana Pina <appinaf@unal.edu.co> and David Zamora <dazamoraa@unal.edu.co> 
Water Resources Engineering Research Group - GIREH
}
\references{
Siena, M., Riva, M., Giamberini, M., & Gouze, P. (2017). Statistical modeling of gas-permeability 
spatial variability along a limestone core. Spatial Statistics.
https://doi.org/10.1016/j.spasta.2017.07.007
}

