% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDFs.R
\name{pwak1}
\alias{pwak1}
\title{Cumulative probability of Wakeby distribution}
\usage{
pwak1(q, xi, alpha, beta, gamma, delta)
}
\arguments{
\item{q:}{a real value vector.}

\item{xi:}{\eqn{\xi} is a location parameter (see \code{\link{parwak}} function).}

\item{alpha:}{\eqn{\alpha} is a location parameter (see \code{\link{parwak}} function).}

\item{beta:}{\eqn{\gamma} is a shape parameter (see \code{\link{parwak}} function).}

\item{gamma:}{\eqn{\gamma} is a shape parameter (see \code{\link{parwak}} function).}

\item{delta:}{\eqn{\delta} is a shape parameter (see \code{\link{parwak}} function).}
}
\value{
Nonexceedance probability (\emph{F}) for q.
}
\description{
This function computes the cumulative probability or nonexceedance 
probability of the Wakeby distribution given parameters (\eqn{\xi}, \eqn{\alpha}, \eqn{\beta}, \eqn{\gamma} and \eqn{\delta}). 
Based in function of \pkg{lmomco} package.
}
\examples{

}
\author{
David Zamora <dazamoraa@unal.edu.co> 
Water Resources Engineering Research Group - GIRE
}
