% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resoPLOT.R
\name{resoPLOT}
\alias{resoPLOT}
\title{Selection of the graphs resolution}
\usage{
resoPLOT(reso = 300)
}
\arguments{
\item{reso}{a number to determine the resolution that the plot function will used to save 
graphs. It has two options: 300 and 600 ppi.}
}
\value{
a vector with four values. They are used 
to save graphs by means of \code{\link{png}} function.

\itemize{
 \item \code{reso}
 \item \code{wth}
 \item \code{hth}
 \item \code{hth1}
}
}
\description{
Selection of the graphs resolution
}
\author{
David Zamora <dazamoraa@unal.edu.co> 
Water Resources Engineering Research Group - GIREH
}
