% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selecDIST.R
\name{selecDIST}
\alias{selecDIST}
\title{Convert large names of distribution functions in a short one}
\usage{
selecDIST(Type = "Gumbel")
}
\arguments{
\item{Type:}{a character specifying the name of the distribution function that 
will be employed: exponential, gamma, gev, gumbel, log.normal3, normal, 
log.pearson3 and wakeby.}
}
\value{
a character with an abbreviation of distribution function defined in \code{Type}.
}
\description{
This function converts large names of distribution functions in a short one 
to used it in functions of \pkg{IDFtool} package.
}
\examples{
(selecDIST(Type = "log.normal3"))
# "ln3

}
\author{
David Zamora <dazamoraa@unal.edu.co> 
Water Resources Engineering Research Group - GIREH
}
