% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Volcanico.R
\docType{data}
\name{Volcanico}
\alias{Volcanico}
\title{Volcanico:  fracture lengths in groundwater bodies measured on outcrops in the La Linea 
massif in the central  range  of  the  Colombian  Andes}
\format{A numeric vector with 468 rows:
\describe{
  \item{Persistencia}{fracture lengths, in meters}   
  ...
}}
\usage{
Volcanico
}
\description{
Fracture lengths measured on outcrops in the La Linea massif in the central  
range  of  the  Colombian  Andes. Outcrops lenght is less than 30 m as the
dense vegetation impeded the location of extensiver areas. Fourteen 
two-dimensional fracture patterns with less than 200 fractures on each 
outcrop were considered.
}
\references{
Hernandez, F., Kammer, A., 2016. Caracterización estructural de los  complejos
Cajamarca y Quebradagrande en la zona del túnel de la línea, con implicaciones
hidrogeologicas. Thesis, Universidad Nacional de Colombia.
}
\keyword{datasets}

