% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IDFCurve.R
\name{IDFCurve}
\alias{IDFCurve}
\title{Compute Intensity-Duration-Frequency curve}
\usage{
IDFCurve(Data = ..., Station = "2610516", Duration = FALSE,
  Periods = FALSE, Type = "gumbel", M.fit = "lmoments", Plot = 1234,
  Strategy = 1:3, logaxe = "", CI = FALSE, CIpdf = TRUE, iter = 500,
  goodtest = FALSE, Resolution = 300, SAVE = FALSE, name = TRUE)
}
\arguments{
\item{Data:}{a numeric matrix with years in the first column and intensity values for each \code{Duration} time in the other columns.}

\item{Station:}{a string with a name to identify source of \code{Data}.}

\item{Duration:}{a logical value or numeric vector. If it is TRUE the durations (in minutes) 
by default will be used: 5, 10 ,15, 20, 30, 60, 120 y 360. In case of a numeric vector the durations must be in minutes.}

\item{Periods:}{a logical value or numeric vector.}

\item{Type:}{a character specifying the name of the distribution function that will 
be employed: exponencial, gamma, gev, gumbel, log.normal3, normal, pearson, log.pearson3 and 
wakeby (see \code{\link{selecDIST}}).}

\item{M.fit:}{a character specifying a name of fit method employed on pdf, just three 
options are available: L-moments (\emph{Lmoments}), Probability-Weighted Moments (\emph{PWD}), 
Maximum Likelihood (\code{\link{MLEZ}}) and Moments (\emph{MME}) (see \code{\link{MME_DIST}}).}

\item{Plot:}{it is a number of one to four digits. a number (1) to plot density curves 
both empirical as modeled (\emph{pdf}).(2) to plot if it will be 
plotted curves between return \code{Periods} and intensity computed by \emph{pdf} fitted. 
Or use (12) to get both graphs. (3) to plot IDF curves for all return periods: (\code{Durations} versus \code{Intensity})
for all return \code{Periods}. (4) to plot IDF curve each for return
\code{Periods} with its confidence and prediction intervals. Or use (34) to get both graphs. If other is used the graphs will not appear.}

\item{Strategy:}{a numeric vector used to identify Strategies to compute IDF curves with different data sets: 1 just data from Ideam, 
2 just data from HIDFUN tool and 3 used this data sets.}

\item{logaxe:}{a character to plot axis in log scale: x, y or both (xy). In other case use "".}

\item{CI:}{a logical value specifying whether confidence and prediction intervals will be computed.}

\item{CIpdf:}{a logical value specifying whether confidence of pdf will be computed.}

\item{iter:}{an integer representing number of resamples to conduct when 
confidence interval is computed (see \code{\link{bootstrapCI}}). Use it only if 
CI is equal to TRUE.}

\item{goodtest:}{a logical value specifying whether goodness-fit tests should be 
cumputed to \emph{pdf} fitted by means of \code{\link{goodfit}} function.}

\item{Resolution:}{a number to determine the resolution that the plot function will used to save graphs. 
It has two options: 300 and 600 ppi. See \code{\link{resoPLOT}}.}

\item{SAVE:}{a logical value. TRUE will save \code{Plot}, FALSE will just show \code{Plot}.}

\item{name:}{a logical value. TRUE will use a default names to identify strategies: (1) "HIDFUN", (2) "IDEAM", (3) "AMBOS". In other case FALSE
allows: i) selected years of data sets, and ii) to insert durations to do IDF curves (in minutes)}
}
\value{
A list of:

 \itemize{
   \item \code{Intesities} a numeric matrix of intensities values per each return \code{Periods} computed by \emph{pdf} fitted.
   \item \code{Models} a list with results of the function \code{\link{regIDF}}.
   \item \code{Test.fit} a list with results of the function \code{\link{goodFIT}}.
   \item \code{Distribution} a list with results of the function \code{\link{fitDISRTI}}.
 }
}
\description{
An Intensity-Duration-Frequency curve (IDF Curve) is a graphical representation 
of the probability that a given average rainfall intensity will occur. This function allows to fit different 
probability distribution functions (see \code{\link{selecDIST}}) by means of four fit methods (see \code{\link{fitDISTRI}})
to determine intensity [mm/h] for different return periods and per specific time durations. Finally, it computes equations parameters 
of the IDF curves (see \code{\link{regIDF}})
}
\examples{

# Meteorology station in the Farfan Airport in Tulua, Colombia.
data(inten)
Test.idftool <- IDFCurve(Data = inten, Station='2610516', Duration = FALSE,
Periods = FALSE, Type = "gumbel", M.fit = "lmoments",
Plot = 1234, Strategy = 1, logaxe = "", CI = FALSE, iter = 100,
goodtest = FALSE, Resolution = 300, SAVE = FALSE, name = TRUE)

}
\author{
David Zamora <dazamoraa@unal.edu.co>
Albeiro Figueroa <cafigueroao@unal.edu.co> 
Water Resources Engineering Research Group - GIREH
}

