% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLEZ.R
\name{MLEZ}
\alias{MLEZ}
\title{Fitting parameters of PDF by means of maximum likelihood}
\usage{
MLEZ(Intensity, type, para.int = NULL, silent = TRUE,
  null.on.not.converge = TRUE, pretransf = function(t) return(t))
}
\arguments{
\item{pretransf}{}

\item{Intensity:}{a numeric vector with intensity [mm/h] values of different
years for a specific time duration (\emph{e.g.} 5, 15, 120 minutes, \emph{etc}.).}

\item{type:}{a character specifying the name of the distribution function that will 
be employed: exponencial, gamma, gev, gumbel, log.normal3, normal, pearson, log.pearson3 and 
wakeby (see \code{\link{selecDIST}}).}

\item{para.int:}{Initial parameters as a vector \eqn{\Theta}.}

\item{silent:}{a logical to silence the \code{\link{try}} function wrapping the \code{\link{DEoptim}} function.}

\item{null.on.not.converge:}{a logical to trigger simple return of NULL if the \code{\link{DEoptim}} function 
returns a nonzero convergence status.#'}
}
\value{
A list of:

 \itemize{
   \item \code{Parameters} a list with type of distribution fitted and values of its parameters
 }
}
\description{
This function fits parameters of probability distribution functions (see \code{\link{selecDIST}})
by means of maximum likelihood and performs evolutionary global optimization via the 
Differential Evolution algorithm (see \pkg{DEoptim} package).
}
\examples{

data(inten) 
TEST.MLE <- MLEZ(Intensity = inten[,4], type = "Gumbel", 
                 para.int = NULL, silent = TRUE, 
                 null.on.not.converge = TRUE, 
                 pretransf = function(t) return(t))
## Results: xi = 71.178 ; alpha = 15.204 

}
