% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_criteria.R
\name{p.criteria}
\alias{p.criteria}
\title{pW Criteria}
\usage{
p.criteria(metrics = .., numod = .., pdfnames = ..)
}
\arguments{
\item{metrics}{}

\item{numod}{}

\item{pdfnames}{}
}
\description{
This function determine what probability distribution function has the best
goodness-of-fit to observations, number of parameters and the quality of 
model parameter estimates, taking in account different criterias information in a 
\emph(Siena et al., 2017).
}

