% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_criteria.R
\name{p.criteria}
\alias{p.criteria}
\title{pW Criteria}
\usage{
p.criteria(metrics = .., critnames = .., pdfnames = ..)
}
\arguments{
\item{metrics:}{a numeric matrix with the values of criterias AIC, BIC, AICc and KIC (rows)
evaluated by each PDFs (columns).}

\item{critnames:}{a character vector with the names of information criteria evaluated.}

\item{pdfnames:}{a character vector with the names of PDFs evaluated.}
}
\value{
Provided a matrix with values of pW criteria by each PDFs and information criterias.
}
\description{
This function determine what probability distribution function has the best
goodness-of-fit to observations, number of parameters and the quality of 
taking in account different information criterias that are evaluated in 
the criteria proposed by (\emph{Siena et al., 2017}).
}
\examples{
# Example of five PDFs and their respectives values of four information criterias.
data(fractures.crit)

pW.1 <- p.criteria(metrics = fractures.crit, critnames = rownames(fractures.crit), pdfnames = colnames(fractures.crit))
pW.1

head(round(pW.1, 2))
# GEV: AIC= 1, BIC= 1, AICc= 1, KIC= 1

}
\author{
Adriana Pina <appinaf@unal.edu.co> and David Zamora <dazamoraa@unal.edu.co> 
Water Resources Engineering Research Group - GIREH
}
\references{

}

